Convert valid_pred_scores to float32 in sim_ota_assigner
Workaround to resolve `RuntimeError: "sqrt" "_vml_cpu" not implemented for 'Half'`

+++ mmdet/core/bbox/assigners/sim_ota_assigner.py
@@ -158,7 +158,7 @@ class SimOTAAssigner(BaseAssigner):
 
         valid_pred_scores = valid_pred_scores.unsqueeze(1).repeat(1, num_gt, 1)
         cls_cost = F.binary_cross_entropy(
-            valid_pred_scores.sqrt_(), gt_onehot_label,
+            valid_pred_scores.to(dtype=torch.float32).sqrt_(), gt_onehot_label,
             reduction='none').sum(-1)
 
         cost_matrix = (
